requirements.txt
certifi==2021.5.30
cycler==0.10.0
decorator==4.4.2
glfw==2.1.0
imageio==2.9.0
kiwisolver==1.3.1
matplotlib==3.4.2
networkx==2.5.1
numpy==1.21.0
opencv-python==4.5.2.54
Pillow==8.2.0
PyOpenGL==3.1.5
PyOpenGL-accelerate==3.1.5
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.18.2
scipy==1.7.0
six==1.16.0
tifffile==2021.6.14
triangle==20200424
